requirements.txt
pandas==2.1.1
scikit_learn==1.3.1
imbalanced-learn==0.11.0
matplotlib==3.8.0
yfinance==0.2.30
jupyter==1.0.0
python-dotenv==1.0.0
# seaborn==0.13.0
